simplify server & csv to json script
