+++ package.json
@@ -5,6 +5,7 @@
   "scripts": {
     "eslint": "./node_modules/.bin/eslint --fix _app",
     "start": "node _app/server.js",
+    "data": "node scripts/csv_to_json.js",
     "dev": "static-server public"
   },
   "dependencies": {
